quote_bot/requirements.txt
beautifulsoup4==4.12.3
lxml==5.3.0
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
